% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.borders.R
\name{grid.borders}
\alias{grid.borders}
\title{Title}
\usage{
grid.borders(
  top = c(l = "", c = "", r = ""),
  bottom = c(l = "", c = "", r = ""),
  left = c(b = "", c = "", t = ""),
  right = c(b = "", c = "", t = ""),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  grid.caption.params = list(),
  watermark = NULL,
  grid.watermark.params = list(),
  clear = FALSE
)
}
\arguments{
\item{clear}{}
}
\description{
Title
}
