% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.borders.R
\name{grid.caption}
\alias{grid.caption}
\title{Title}
\usage{
grid.caption(
  label = "Notes & Sources:",
  caption = c(),
  y = NULL,
  nudge_x = 0,
  nudge_y = NULL,
  sep = 0.02,
  cex = 5/6
)
}
\arguments{
\item{cex}{}
}
\description{
Title
}
