% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.borders.R
\name{grid.slip}
\alias{grid.slip}
\title{Title}
\usage{
grid.slip(
  label = "",
  draw = TRUE,
  layout = "portrait",
  fontsize = 12,
  cex = 2,
  fontfamily = "LM Roman 10",
  fontface = "bold"
)
}
\arguments{
\item{fontface}{}
}
\description{
Title
}
