% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.table.minimal.R
\name{grid.table.gp}
\alias{grid.table.gp}
\title{Title}
\usage{
grid.table.gp(
  g,
  cols = NULL,
  rows = NULL,
  union = TRUE,
  name = "core-fg",
  bold = F,
  gp_params = list()
)
}
\arguments{
\item{gp_params}{}
}
\description{
Title
}
