% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.table.minimal.R
\name{grid.table.minimal}
\alias{grid.table.minimal}
\title{Title}
\usage{
grid.table.minimal(
  DT,
  x = 0.5,
  y = 0.45,
  fontsize = NULL,
  booklines = TRUE,
  hlines = NULL,
  hline = gpar(col = "black", lwd = 1),
  padding = unit(c(0, 0), "npc"),
  widths = NULL,
  heights = NULL,
  width_pct = 0.75,
  height_pct = 0.5,
  black.heading = FALSE,
  rows = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
