% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.borders.R
\name{grid.watermark}
\alias{grid.watermark}
\title{Title}
\usage{
grid.watermark(
  label = "",
  draw = TRUE,
  x = 0.5,
  y = 0.5,
  rot = 45,
  fontsize = 12,
  cex = 8,
  alpha = 0.1,
  color = "grey",
  fontface = "bold"
)
}
\arguments{
\item{fontface}{}
}
\description{
Title
}
