% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.table.minimal.R
\name{gtable_add_cspan}
\alias{gtable_add_cspan}
\title{Title}
\usage{
gtable_add_cspan(
  g,
  text = "",
  t = 0,
  b = 0,
  l = 1,
  r = NULL,
  black = FALSE,
  text.color = "black",
  just = c("center", "center"),
  fontface = "bold",
  line = TRUE,
  line.color = "black",
  line.lwd = 1,
  line.sep = 0.1,
  bg_params = NULL,
  height = NULL
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
