% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.table.minimal.R
\name{gtable_add_hline}
\alias{gtable_add_hline}
\title{Title}
\usage{
gtable_add_hline(
  g,
  rows,
  gp = gpar(col = "black", lwd = 1),
  l = 1,
  r = ncol(g),
  pos = 0
)
}
\arguments{
\item{pos}{}
}
\description{
Title
}
