% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_border_defaults.R
\name{set_border_defaults}
\alias{set_border_defaults}
\title{Title}
\usage{
set_border_defaults(
  show = FALSE,
  plot.width = 11,
  plot.height = 8.5,
  plot.sep = 0.02,
  border.margin = 0.05,
  caption.margin = 0.05,
  border.radius = 0.009,
  title.margin = 0.05,
  title.radius = 0.1,
  title.width = 0.8,
  title.height = 0.1,
  title.sep = 0.011,
  fontsize = 12,
  border.fontsize = 12,
  default.fontfamily = "LM Roman 10",
  border.fontfamily = "CMU Sans Serif",
  title.fontface = "bold",
  subtitle.fontface = "italic"
)
}
\arguments{
\item{subtitle.fontface}{}
}
\description{
Title
}
