% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_borders.R
\name{theme_borders}
\alias{theme_borders}
\title{Title}
\usage{
theme_borders(
  axis.text = element_text(color = "black"),
  axis.ticks = element_line(),
  axis.line = element_line(linewidth = 0.5, color = "black"),
  panel.grid.major = element_blank(),
  panel.grid.minor = element_blank(),
  legend.justification.inside = c(1, 1),
  legend.text = element_text(face = "plain", size = rel(4/6)),
  legend.title = element_text(face = "plain", size = rel(5/6)),
  legend.margin = margin(t = 0, b = 0, l = 0, r = 0, unit = "in"),
  legend.box.margin = margin(t = 0, b = 0, l = 0, r = 0, unit = "in"),
  strip.text = element_text(face = "plain", size = rel(1)),
  plot.title = element_blank(),
  plot.subtitle = element_blank(),
  plot.caption = element_blank(),
  plot.caption.position = "panel",
  plot.title.position = "panel",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
